[minor] fix compiler warnings in OP2 MPI library

+++ op2/c/src/externlib/op_hdf5.c
@@ -190,7 +190,7 @@ op_dat op_decl_dat_hdf5(op_set set, int dim, char const *type, char const *file,
     file_id = H5Fopen(file, H5F_ACC_RDONLY, H5P_DEFAULT);
     
     /*find element size of this dat with available attributes*/
-    int dat_size = 0;
+    size_t dat_size = 0;
     //open existing data set
     dset_id = H5Dopen(file_id, name, H5P_DEFAULT);
     //get OID of the attribute
@@ -249,7 +249,7 @@ op_dat op_decl_dat_hdf5(op_set set, int dim, char const *type, char const *file,
     	
     	if(dat_size != dim*sizeof(double))
     	{
-    	    printf("dat.size %d in file %s and %d*sizeof(double) do not match\n",dat_size,file,dim);
+    	    printf("dat.size %lu in file %s and %d*sizeof(double) do not match\n",dat_size,file,dim);
     	    exit(2);
     	}
     	else
@@ -262,7 +262,7 @@ op_dat op_decl_dat_hdf5(op_set set, int dim, char const *type, char const *file,
     	
     	if(dat_size != dim*sizeof(float)) 
     	{
-    	    printf("dat.size %d in file %s and %d*sizeof(float) do not match\n",dat_size,file,dim);
+    	    printf("dat.size %lu in file %s and %d*sizeof(float) do not match\n",dat_size,file,dim);
     	    exit(2);
     	}
     	else
@@ -275,7 +275,7 @@ op_dat op_decl_dat_hdf5(op_set set, int dim, char const *type, char const *file,
     	
     	if(dat_size != dim*sizeof(int)) 
     	{
-    	    printf("dat.size %d in file %s and %d*sizeof(int) do not match\n",dat_size,file,dim);
+    	    printf("dat.size %lu in file %s and %d*sizeof(int) do not match\n",dat_size,file,dim);
     	    exit(2);
     	}
     	else
